Added attribute listener on 'commands.next', so you can get notified when a waypoint is reached and the drone starts heading for the next one.
